+++ packages/daemon-images/lib/index.js
@@ -1,5 +1,5 @@
 const fs = require('fs')
-const move = require('./lib/move')
+const move = require('./move')
 
 module.exports = {
   // Used primarily for testing. Or if you want to inject a function to do your reading for you.
